+++ CHANGELOG.md
@@ -21,6 +21,8 @@ Categories Used:
 
 ### New Features
 
+- List: show symlink targets (for tar and zip) (https://github.com/ouch-org/ouch/pull/934)
+
 ### Improvements
 
 ### Bug Fixes